+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 
 ## 2.10.3
 
+### New Features
+
+- **Amazon Transcribe Streaming**
+  - Amazon Transcribe streaming transcription enables you to send an audio stream and receive a stream of text in real time using WebSockets.
+
 ### Bug Fixes
 
 - **AWS Core**